+++ ios/universal_bluetooth/Sources/universal_bluetooth/ExternalAccessory.g.swift
@@ -215,8 +215,7 @@ class ExternalAccessoryChannelSetup {
 ///
 /// Generated protocol from Pigeon that represents Flutter messages that can be called from Swift.
 protocol ExternalAccessoryCallbackChannelProtocol {
-  func accessoryConnected(accessory accessoryArg: EAAccessory, completion: @escaping (Result<Void, PigeonError>) -> Void)
-  func accessoryDisconnected(accessory accessoryArg: EAAccessory, completion: @escaping (Result<Void, PigeonError>) -> Void)
+  func onConnectionStateChanged(accessory accessoryArg: EAAccessory, connected connectedArg: Bool, completion: @escaping (Result<Void, PigeonError>) -> Void)
 }
 class ExternalAccessoryCallbackChannel: ExternalAccessoryCallbackChannelProtocol {
   private let binaryMessenger: FlutterBinaryMessenger
@@ -228,28 +227,10 @@ class ExternalAccessoryCallbackChannel: ExternalAccessoryCallbackChannelProtocol
   var codec: ExternalAccessoryPigeonCodec {
     return ExternalAccessoryPigeonCodec.shared
   }
-  func accessoryConnected(accessory accessoryArg: EAAccessory, completion: @escaping (Result<Void, PigeonError>) -> Void) {
-    let channelName: String = "dev.flutter.pigeon.universal_bluetooth.ExternalAccessoryCallbackChannel.accessoryConnected\(messageChannelSuffix)"
+  func onConnectionStateChanged(accessory accessoryArg: EAAccessory, connected connectedArg: Bool, completion: @escaping (Result<Void, PigeonError>) -> Void) {
+    let channelName: String = "dev.flutter.pigeon.universal_bluetooth.ExternalAccessoryCallbackChannel.onConnectionStateChanged\(messageChannelSuffix)"
     let channel = FlutterBasicMessageChannel(name: channelName, binaryMessenger: binaryMessenger, codec: codec)
-    channel.sendMessage([accessoryArg] as [Any?]) { response in
-      guard let listResponse = response as? [Any?] else {
-        completion(.failure(createConnectionError(withChannelName: channelName)))
-        return
-      }
-      if listResponse.count > 1 {
-        let code: String = listResponse[0] as! String
-        let message: String? = nilOrValue(listResponse[1])
-        let details: String? = nilOrValue(listResponse[2])
-        completion(.failure(PigeonError(code: code, message: message, details: details)))
-      } else {
-        completion(.success(Void()))
-      }
-    }
-  }
-  func accessoryDisconnected(accessory accessoryArg: EAAccessory, completion: @escaping (Result<Void, PigeonError>) -> Void) {
-    let channelName: String = "dev.flutter.pigeon.universal_bluetooth.ExternalAccessoryCallbackChannel.accessoryDisconnected\(messageChannelSuffix)"
-    let channel = FlutterBasicMessageChannel(name: channelName, binaryMessenger: binaryMessenger, codec: codec)
-    channel.sendMessage([accessoryArg] as [Any?]) { response in
+    channel.sendMessage([accessoryArg, connectedArg] as [Any?]) { response in
       guard let listResponse = response as? [Any?] else {
         completion(.failure(createConnectionError(withChannelName: channelName)))
         return